+++ src/main/java/com/ptithcm/apt/repository/ApartmentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ptithcm.apt.entity.Apartment;
 import java.util.List;
 
-public interface ApartmentRepository extends JpaRepository<Apartment, Integer> {
+public interface ApartmentRepository extends JpaRepository<Apartment, Long> {
     boolean existsByRoomNumber(String roomNumber);
 
     List<Apartment> findByStatus(String status);
